Move config to its own file.
Future proof against multiple utils
